requirements.txt

flask
flask-cors
google-generativeai
langchain
sentence-transformers
faiss-cpu
torch
transformers
requests
python-dotenv